Move sourcecode classes
The HTML elements created for a `<sourcecode>` element have class
attributes attached to the `<pre>` element.  In comparison, `<artwork>`
is rendered with class attritributes attached to the outer `<div>`.

Having the classes on the outer `<div>` makes it a LOT easier to style
these properly, so I've moved them in this CL.