feat: Remove warning about SVG width and height (#1225)
This warning annoys a lot of people unnecessarily.

The argument that this means that the diagram doesn't stretch is true
(it won't stretch), but misleading, because stretching is not a good
goal to seek.  Diagrams in RFCs often have text, which either becomes too large
or too small when stretched.  Consequently, the right choice is to set
width and height in many cases.

Indeed, the example in RFC 7992 includes width and height (and no
viewbox, but that's a separate issue).  RFC 7996 does nothing about
this.  This warning is therefore overreach.